Add back the `-l` to trayer.
This is super tricky. At the end of the day, we want the following
layering (from highest z index to lowest, aka top to bottom):

    +-----------------+
    | regular windows |
    +-----------------+
    | struts - trayer |
    +-----------------+
    | struts - xmobar |
    +-----------------+

Normally the realationship between "regular windows" and
our "struts" (trayer+xmobar) doesn't matter, because our windows are
positioned such that the won't overlap with our struts. However, when
doing a mod+b, our windows are supposed to ignore struts, and that means
they should end up on top of both xmobar and trayer.

We want trayer to be on top of xmobar because we've actually configured
xmobar to be 100% wide and to dynamically allocate empty space for where
trayer should be. See
[trayer-padding-icon.sh](https://github.com/jfly/dotfiles/blob/master/bin/trayer-padding-icon.sh)
for more information about this janky hack.

Here's what I think `size_tray` function previously did:

1. It restarts xmonad. This kills xmobar, but xmobar does not
immediately start up.
2. It restarts trayer. Note that we
were *not* having trayer lower itself on start here, so this results in
trayer on the very top.
3. xmobar finishes starting up, and lowers itself to the bottom of the
window stack with its default config value of `lowerOnStart=True`. Note
that this gets xmobar below trayer (as desired), but only works because
it took a moment for xmobar to boot. This is kind of racy.

The above sequence of events leads to the following z stack:

    +-----------------+
    | struts - trayer |
    +-----------------+
    | regular windows |
    +-----------------+
    | struts - xmobar |
    +-----------------+

This meant if you toggle struts, then trayer will show up on top of
regular windows. No good! I changed it to instead *first* start and
lower trayer, and *then* start and lower xmobar. This should reliably
get us into the desired z stack, without relying upon xmobar taking a
moment to boot up.