Finally configure docker detachKeys!
This gets ctrl+p and ctrl+n working as I'd expect when inside of a
container.

This was so darn tricky to do because by default, docker stores
credentials in the `.docker/config.json` file. You have to configure a
`credsStore` to avoid that. Dumping all the tabs I have open:

  - https://docs.docker.com/engine/reference/commandline/login/#credential-helpers
  - https://aur.archlinux.org/packages/docker-credential-secretservice/
  - https://www.antoniojgutierrez.com/2018/08/11/docker_login_password_store.html
  - https://github.com/docker/docker-credential-helpers