Cleaning up/refactoring my asdf+direnv installation.
This is a mess: the most important bit is the unsetting of *all* the
`DIRENV_*` environment variables. I was previously only unsetting
some of them, which would leave direnv in a weird state with its
watches, and cause a strange issue where asdf-direnv would regenerate
its cached env var in a different way.

Probably the *right* fix is to instead do the PATH setup stuff once in
`~/.zprofile`, and then any PATH and DIRENV_* env vars will just
automagically percolate down into subshells, and we'll save a little
time too (by not having to re-run `direnv`). But, that doesn't quite
work with the way I like to use `shtuff` inside of my folders: my
`shtuff` command itself munges the `PATH`!

    cat $(which shtuff)
    #! /nix/store/07j81a5xy8j1srvf57dl7lxsaihnmfac-bash-5.1-p12/bin/bash -e
    export PATH='/nix/store/x24kxyqwqg2ln8kh9ky342kdcmhbng3h-python3-3.9.9/bin:/nix/store/a4k7wb3g8cdmxzjnwcp13jhw1msvigl1-shtuff-0.3.1/bin'${PATH:+':'}$PATH
    export PYTHONNOUSERSITE='true'
    exec -a "$0" "/nix/store/a4k7wb3g8cdmxzjnwcp13jhw1msvigl1-shtuff-0.3.1/bin/.shtuff-wrapped"  "$@"

This is actually quite nice: it means we're not subject to whatever
version of python we happen to be using in whatever repo I'm working in,
but it would be nice if this could work somehow without actually munging
the PATH. I think this is a nix question...