Alternative approach to managing my PATH
Now it only ever gets set when loading up a "login shell", which should
be pretty rare. This means it should be possible to start a subshell
inside of a direnv managed directory, and the `PATH` shouldn't change,
and direnv shouldn't get confused.

It's just a little unfortunate that my setup has to be split across a
zprofile and zshrc files. I don't think that's going to be super easy to
explain to people.

This also required hacking on `shtuff` a bit to make sure that *it*
doesn't munge the `PATH` at all (not really a shtuff thing, it's
actually a nix thing). Hopefully there's a more correct way of doing
this.