Add a script for concatenating pdfs
Unfortunately, `convert *.jpg out.pdf` didn't quite work for me. It
seems like it should, but the last time I tried to do it, I ended up
with pages of inconsistent sizes.

This was tricky enough to get working that I wanted to have a script to
remind myself how to do this.